+++ ios/flutter_mapbox_navigation_plus/Sources/flutter_mapbox_navigation_plus/models/Feedback.swift
@@ -1,10 +1,3 @@
-//
-//  Feedback.swift
-//  flutter_mapbox_navigation
-//
-//  Created by Marco Galetta on 20/05/23.
-//
-
 import Foundation
 
 public class Feedback : Codable
@@ -12,7 +5,7 @@ public class Feedback : Codable
     let rating: Int?
     let comment: String?
 
-    init(rating: Int?,comment: String?) {
+    init(rating: Int?, comment: String?) {
         self.rating = rating
         self.comment = comment
     }